[OC-11574] Do not interact with keepalived by default.
In HA topologies, keepalived is responsible for managing failover
between the machines.  If the keepalived service is stopped, a
failover occurs.  Typically, when I user runs `private-chef-ctl
restart`, they do not expect a failover to occur.

This patch ensures that we only modify the state of the keepalived
service when explicitly told to.