blacklist removed services from *-ctl commands
When running omnibus-ctl [status|start|restart|etc] for services
that have been removed and are configured as removed (requires
config to be written by the *-omnibus projects), we should not
attempt to start the services or run any service commands against
them unless specifically attempting to do so. Services that are
removed can still be accessed by running:

```
*-ctl SV_COMMAND REMOVED_SERVICE
```

The previous behavior manifested itself as a race condition in
Enterprise Chef where a service that was previously installed as
part of an older version came up after a `*-ctl start` and stole
the port from the service that was supposed to be running. This
fixes that issue.