Remove old node state before running reconfigure or show-config.
Currently, the reconfigure cookbooks for most projects using
omnibus-ctl include complex option processing where the final
configuration is merged into the node data.  Chef-client -z loads the
node state from the last run, which leads to situations in which a
user can remove configuration from the main configuration file but the
configuration isn't actually reverted to the default when reconfigure
is run.

In the long run we want to move away from overloading node attribtues
in this way. However, to avoid massive changes to the configuration
of a number of projects, we simply nuke the old node data.

Fixes chef/opscode-omnibus#730
Fixes chef/omnibus-ctl#24