Directly use the embedded sv instead of the init symlinks.
The symlinks in init/ are created by the runit cookbook, and on Debian
it creates a wrapper script instead of simple symlinks. This wrapper
script encodes `/usr/bin/sv` as the control binary.

Instead of relying on the symlinks being as expected, directly use the
binary that is embedded anyway.

Fixes #53. Fixes running chef-server on Debian.

Signed-off-by: Matthias Rampke <mr@soundcloud.com>