Explicitly set SVDIR to the service_dir
We currently have a bug on Debian where we rely on the
system-installed runit. The system installed runit does not have our
patch to reset the servicedir. Users were previously working around
this by setting SVDIR explicitly. This preserves those users ability
to use Chef server on the next release. The full fix will be to ensure
we don't rely on the system-installed runit on Debian.

Signed-off-by: Steven Danna <steve@chef.io>